# specs/001-proposal-assistant/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: EasyGrant Smart Proposal Assistant API
  description: |
    REST API for the Smart Proposal Assistant demo. Supports funding call upload,
    requirements extraction, RAG-based section generation with citations, sticky edits,
    and DOCX export.
  version: 1.0.0
  contact:
    name: EasyGrant Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://easygrant.onrender.com
    description: Production (Render)
...
paths:
  /api/upload/funding-call:
    post:
      tags: [Upload]
      summary: Upload funding call PDF
      description: |
        Upload a funding opportunity PDF. System extracts requirements (sections,
        word limits, scoring criteria) and returns a structured blueprint.
        Max file size: 10MB. (FR-001)
      operationId: uploadFundingCall
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file of funding call
                session_id:
                  type: string
                  format: uuid
                  description: Optional session ID (created if not provided)
              required: [file]
      responses:
        '200':
          description: Upload successful, requirements extracted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingCallResponse'
        '400':
          description: Invalid file (not PDF, >10MB, or parsing failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error during extraction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/upload/supporting-docs:
    post:
      tags: [Upload]
      summary: Upload supporting documents (local context)
      description: |
        Upload up to 5 supporting documents (PDF, DOCX, TXT) containing local
        community context. System chunks and indexes into vector store. Total
        size limit: 50MB across all docs. (FR-002, FR-004)
      operationId: uploadSupportingDocs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
                  description: Up to 5 PDF/DOCX/TXT files
                session_id:
                  type: string
                  format: uuid
              required: [files, session_id]
      responses:
        '200':
          description: Upload and indexing successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportingDocsResponse'
        '400':
          description: Invalid files (>5 docs, >50MB total, unsupported format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    FundingCallResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        funding_call_id:
          type: string
          format: uuid
        filename:
          type: string
          example: "Community_Grant_RFP_2025.pdf"
        total_pages:
          type: integer
          example: 12
        blueprint:
          $ref: '#/components/schemas/RequirementsBlueprint'
        extraction_status:
          type: string
          enum: [pending, processing, completed, failed]
          example: completed
      required: [session_id, funding_call_id, filename, blueprint, extraction_status]
    RequirementsBlueprint:
      type: object
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionRequirement'
        eligibility_criteria:
          type: array
          items:
            type: string
          example: ["Serve population <5000", "Non-profit organization"]
        deadline:
          type: string
          format: date
          example: "2025-12-31"
        funding_amount:
          type: string
          example: "$50,000 - $100,000"
        scoring_criteria:
          type: object
          additionalProperties:
            type: integer
          example:
            need: 20
            approach: 30
            capacity: 25
            budget: 25
    SectionRequirement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Community Need Statement"
        order:
          type: integer
          example: 1
        word_limit:
          type: integer
          nullable: true
          example: 500
        char_limit:
          type: integer
          nullable: true
        format_requirements:
          type: string
          nullable: true
          example: "Narrative paragraph"
        scoring_weight:
          type: integer
          nullable: true
          example: 20
        generation_status:
          type: string
          enum: [not_started, generating, completed, failed]
          example: not_started
    SupportingDocsResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        documents:
          type: array
          items:
            $ref: '#/components/schemas/SupportingDocument'
        total_upload_size_mb:
          type: number
          format: float
          example: 24.5
        total_chunks_indexed:
          type: integer
          example: 287
    SupportingDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          example: "Strategic_Plan_2024.pdf"
        file_type:
          type: string
          enum: [pdf, docx, txt]
        file_size_mb:
          type: number
          format: float
        total_pages:
          type: integer
          nullable: true
        chunk_count:
          type: integer
        indexed_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "File size exceeds 10MB limit"
        code:
          type: string
          example: "FILE_TOO_LARGE"
        details:
          type: object
          nullable: true